Moved FdbFuture<T> to the FoundationDb.Client.Native namespace since it is a concept specific to fdb_c.dll
