+++ FoundationDB.Client/Native/FdbFuture.cs
@@ -1,5 +1,5 @@
 ﻿#region BSD Licence
-/* Copyright (c) 2013, Doxense SARL
+/* Copyright (c) 2013-2015, Doxense SAS
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -29,9 +29,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // enable this to help debug Futures
 #undef DEBUG_FUTURES
 
-namespace FoundationDB.Client
+namespace FoundationDB.Client.Native
 {
-	using FoundationDB.Client.Native;
 	using FoundationDB.Client.Utils;
 	using JetBrains.Annotations;
 	using System;
