+++ FoundationDB.Client/Native/FdbFutureSingle.cs
@@ -1,5 +1,5 @@
 ﻿#region BSD Licence
-/* Copyright (c) 2013-2014, Doxense SAS
+/* Copyright (c) 2013-2015, Doxense SAS
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -28,9 +28,8 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 #undef DEBUG_FUTURES
 
-namespace FoundationDB.Client
+namespace FoundationDB.Client.Native
 {
-	using FoundationDB.Client.Native;
 	using JetBrains.Annotations;
 	using System;
 	using System.Diagnostics;
