+++ FoundationDB.Client/Native/Handles/DatabaseHandle.cs
@@ -1,5 +1,5 @@
 ﻿#region BSD Licence
-/* Copyright (c) 2013-2014, Doxense SAS
+/* Copyright (c) 2013-2015, Doxense SAS
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -35,28 +35,28 @@ namespace FoundationDB.Client.Native
 #endif
 	using System.Threading;
 
-	/// <summary>Wrapper on a FDBFuture*</summary>
+	/// <summary>Wrapper on a FDBDatabase*</summary>
 #if MONO
 	[StructLayout(LayoutKind.Sequential)]
 #endif
-	internal sealed class FutureHandle : FdbSafeHandle
+	internal sealed class DatabaseHandle : FdbSafeHandle
 	{
 
-		public FutureHandle()
+		public DatabaseHandle()
 		{
-			Interlocked.Increment(ref DebugCounters.FutureHandlesTotal);
-			Interlocked.Increment(ref DebugCounters.FutureHandles);
+			Interlocked.Increment(ref DebugCounters.DatabaseHandlesTotal);
+			Interlocked.Increment(ref DebugCounters.DatabaseHandles);
 		}
 
 		protected override void Destroy(IntPtr handle)
 		{
-			FdbNative.FutureDestroy(handle);
-			Interlocked.Decrement(ref DebugCounters.FutureHandles);
+			FdbNative.DatabaseDestroy(handle);
+			Interlocked.Decrement(ref DebugCounters.DatabaseHandles);
 		}
 
 		public override string ToString()
 		{
-			return "FutureHandle[0x" + this.Handle.ToString("x") + "]";
+			return "DatabaseHandle[0x" + this.Handle.ToString("x") + "]";
 		}
 
 	}
