+++ FoundationDB.Client/Native/Handles/TransactionHandle.cs
@@ -1,5 +1,5 @@
 ﻿#region BSD Licence
-/* Copyright (c) 2013, Doxense SARL
+/* Copyright (c) 2013-2015, Doxense SAS
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
@@ -35,28 +35,28 @@ namespace FoundationDB.Client.Native
 #endif
 	using System.Threading;
 
-	/// <summary>Wrapper on a FDBCluster*</summary>
+	/// <summary>Wrapper on a FDBTransaction*</summary>
 #if MONO
 	[StructLayout(LayoutKind.Sequential)]
 #endif
-	internal sealed class ClusterHandle : FdbSafeHandle
+	internal sealed class TransactionHandle : FdbSafeHandle
 	{
 
-		public ClusterHandle()
+		public TransactionHandle()
 		{
-			Interlocked.Increment(ref DebugCounters.ClusterHandlesTotal);
-			Interlocked.Increment(ref DebugCounters.ClusterHandles);
+			Interlocked.Increment(ref DebugCounters.TransactionHandlesTotal);
+			Interlocked.Increment(ref DebugCounters.TransactionHandles);
 		}
 
 		protected override void Destroy(IntPtr handle)
 		{
-			FdbNative.ClusterDestroy(handle);
-			Interlocked.Decrement(ref DebugCounters.ClusterHandles);
+			FdbNative.TransactionDestroy(handle);
+			Interlocked.Decrement(ref DebugCounters.TransactionHandles);
 		}
 
 		public override string ToString()
 		{
-			return "ClusterHandle[0x" + this.Handle.ToString("x") + "]";
+			return "TransactionHandle[0x" + this.Handle.ToString("x") + "]";
 		}
 
 	}
